Lots of cleanup and consolidation around search endpoint tests
