+++ api-search/src/integrationTest/groovy/org/cedar/onestop/api/search/IntegrationTestConfig.groovy
@@ -40,57 +40,4 @@ class IntegrationTestConfig {
     ).build()
     return restClient
   }
-
-  @Bean
-  Map testData() {
-    return [
-        'DEM': [
-            'C1': [
-                id: 'e7a36e60-1bcb-47b1-ac0d-3c2a2a743f9b',
-                granules: [],
-                flattenedGranules: []
-            ],
-            'C2': [
-                id: 'e5820283-3686-44d0-8edd-28a086eb500e',
-                granules: [],
-                flattenedGranules: []
-            ],
-            'C3': [
-                id: '1415b3db-c602-4dbb-a502-4091fe9df1cf',
-                granules: [],
-                flattenedGranules: []
-            ],
-        ],
-        'GHRSST': [
-            'C1': [
-                id: '920d8155-f764-4777-b7e5-14442b7275b8',
-                granules: [],
-                flattenedGranules: []
-            ],
-            'C2': [
-                id: '882511bc-e99e-4597-b634-47a59ddf9fda',
-                granules: [],
-                flattenedGranules: []
-            ],
-            'C3': [
-                id: '42ea683d-e4e7-434c-8823-abff32e00f34',
-                granules: [],
-                flattenedGranules: []
-            ],
-        ],
-        'COOPS': [
-            'C1': [
-                id: 'fcf83ec9-964b-45b9-befe-378ea6ce52cb',
-                granules: [
-                    'G1': [id: '783089c4-3484-4f70-ac8d-d4818d0cd0dd'],
-                    'G2': [id: 'a207b48f-29fc-4d79-a676-1f265cd9971f'],
-                ],
-                flattenedGranules: [
-                    'FG1': [id: '783089c4-3484-4f70-ac8d-d4818d0cd0dd'],
-                    'FG2': [id: 'a207b48f-29fc-4d79-a676-1f265cd9971f']
-                ]
-            ]
-        ]
-    ]
-  }
 }
