+++ api-search/src/integrationTest/resources/data/GenericFilterData/geometry_testing.txt
@@ -3,7 +3,7 @@
 { "index": { "_index": "geometry_testing", "_type": "doc", "_id": "2" } }
 { "spatialBounding": { "coordinates": [-76.315, 36.977], "type": "Point" }, "isGlobal": false }
 { "index": { "_index": "geometry_testing", "_type": "doc", "_id": "3" } }
-{ "spatialBounding": { "coordinates": [[[-180, -90], [-175.78, 51.25], [-175.78, 52.16], [-177.24, 52.16], [-177.24, 51.25]]], "type": "Polygon"}, "isGlobal": false }
+{ "spatialBounding": { "coordinates": [[[-177.24, 51.25], [-175.78, 51.25], [-175.78, 52.16], [-177.24, 52.16], [-177.24, 51.25]]], "type": "Polygon"}, "isGlobal": false }
 { "index": { "_index": "geometry_testing", "_type": "doc", "_id": "4" } }
 { "spatialBounding": { "coordinates": [150, 60], "type": "Point" }, "isGlobal": false }
 { "index": { "_index": "geometry_testing", "_type": "doc", "_id": "5" } }
